requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dill==0.3.1.1
future==0.18.2
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
matplotlib==3.1.2
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.2.1
pandas==1.1.4
promise==2.3
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.4.1
six==1.14.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-gpu==2.2.0
tensorflow-datasets==3.1.0
tensorflow-estimator==2.0.1
tensorflow-metadata==0.22.0
termcolor==1.1.0
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
